refactor(ticketing): extract cascade-grouping helpers to recover MI baseline
