+++ baselines/maintainability.json
@@ -131,6 +131,7 @@
   ".agents/scripts/lib/observability/signals-writer.js": 88.188,
   ".agents/scripts/lib/observability/tool-trace-hook.js": 94.506,
   ".agents/scripts/lib/orchestration/automerge-predicate.js": 95.509,
+  ".agents/scripts/lib/orchestration/cascade-grouping.js": 97.534,
   ".agents/scripts/lib/orchestration/code-review.js": 90.238,
   ".agents/scripts/lib/orchestration/concurrency.js": 106.379,
   ".agents/scripts/lib/orchestration/concurrent-task-resolver.js": 89.513,
@@ -206,7 +207,7 @@
   ".agents/scripts/lib/orchestration/task-fetcher.js": 114.391,
   ".agents/scripts/lib/orchestration/ticket-validator-sizing.js": 94.978,
   ".agents/scripts/lib/orchestration/ticket-validator.js": 93.264,
-  ".agents/scripts/lib/orchestration/ticketing.js": 91.261,
+  ".agents/scripts/lib/orchestration/ticketing.js": 92.68,
   ".agents/scripts/lib/orchestration/wave-dispatcher.js": 88.783,
   ".agents/scripts/lib/orchestration/wave-marker.js": 117.783,
   ".agents/scripts/lib/path-security.js": 117.206,
